requirements.txt
# Core dependencies
ollama>=0.3.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
python-frontmatter>=1.0.0
PyYAML>=6.0
python-slugify>=8.0.0

# Langchain (optional for advanced RAG)
langchain>=0.1.0
langchain-community>=0.0.10

# Utilities
rich>=13.7.0
click>=8.1.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Async and HTTP
httpx>=0.25.0
aiohttp>=3.9.0
feedparser>=6.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0

# Optional: Qdrant (alternative vector DB)
qdrant-client>=1.7.0
